Add "desc" parameter to glossary_atoz shortcode
The default is desc = "tip", which matches the behavior before. Additional parameters desc="excerpt" and desc="full" are functional and match the behavior for the glossary_term_list shortcode. However, the parameter desc = "none" is not working--not sure why.

Also, added some html tags to entry header and entry content to make it easier to format the page.